+++ test/cases/html-plugin/webpack.config.js
@@ -7,11 +7,10 @@ module.exports = {
         new HtmlWebpackPlugin({ template: './index.ejs' }),
         new ConcatPlugin({
             uglify: true,
-            useHash: true,
             sourceMap: true,
             name: 'file',
             fileName: '[name].[hash].js',
-            filesToConcat: ['./test/fixtures/a.js', './test/fixtures/b.js']
+            filesToConcat: ['../../fixtures/a.js', '../../fixtures/b.js']
         }),
     ],
 };