+++ test/cases/source-map-use-hash/webpack.config.js
@@ -6,10 +6,9 @@ module.exports = {
     plugins: [
         new ConcatPlugin({
             sourceMap: true,
-            useHash: true,
             name: 'file',
             fileName: '[name].[hash].js',
-            filesToConcat: ['./test/fixtures/a.js', './test/fixtures/b.js']
+            filesToConcat: ['../../fixtures/a.js', '../../fixtures/b.js']
         })
     ]
 };