+++ test/cases/use-hash/webpack.config.js
@@ -5,10 +5,9 @@ module.exports = {
     entry: './index.js',
     plugins: [
         new ConcatPlugin({
-            useHash: true,
             name: 'file',
             fileName: '[name].[hash].js',
-            filesToConcat: ['./test/fixtures/a.js', './test/fixtures/b.js']
+            filesToConcat: ['../../fixtures/a.js', '../../fixtures/b.js']
         })
     ]
 };